Refactor ft_printhex.c to handle negative numbers in 'ft_printhex' function

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: sfarren <sfarren@student.42malaga.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/02 19:20:31 by sfarren           #+#    #+#             */
-/*   Updated: 2024/07/15 09:34:30 by sfarren          ###   ########.fr       */
+/*   Updated: 2024/07/15 15:49:43 by sfarren          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,13 +22,13 @@
 #include <stdio.h>
 
 int		ft_printf(const char *str, ...);
-int		ft_putnbr_base(unsigned long long n, char *base);
+int		ft_putnbr_base(unsigned int n, char *base);
 char	*ft_utoa(unsigned int n);
 int		ft_printchr(char c);
 int		ft_printnbr(int n);
 int		ft_printstr(char *s);
 int		ft_printptr(void *ptr);
-int 	ft_printhex(unsigned long long n, char c);
+int 	ft_printhex(int n, char c);
 int		ft_printu(unsigned int n);
 int		parse_fstring(const char *str, va_list args);
 // int		get_next(const char *str, int i);
